requirements.txt
aioredis==1.3.1
elasticsearch[async]==7.15.1
fastapi==0.70.0
gunicorn==20.1.0
orjson==3.6.4
pip==21.3
pydantic==1.8.2
python-dotenv==0.19.0
psycopg2-binary==2.9.1
uvicorn==0.15.0
uvloop==0.16.0